🐛 fix(registerCommand): Prevent linking when no .env is set
Prior to this change, if anyone except the package author installed any
package that contains @cajacko/template it would ask them for the paths
to the local @cajacko packages. Which they wouldn't have, as they're not
developing them. This exists because there was a check that looks for a
.env file and if it does not exist, tries to use the local packages.

This change only does that check if we're on the package authors machine.

- Original Branch: feature/react-native-eject